creates JSON index route for items
