+++ app/controllers/items_controller.rb
@@ -1,0 +1,7 @@
+class ItemsController < ApplicationController
+
+  def index
+    render json: Item.where(bill_id: params[:bill_id])
+  end
+
+end
