+++ spec/factories/bills.rb
@@ -1,5 +1,5 @@
-FactoryGirl.define do
-  factory :bill do
-    event "party"
-  end
-end
+# FactoryGirl.define do
+#   factory :bill do
+#     event "party"
+#   end
+# end
