+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require 'factory_girl_rails'
+require 'rspec/json_expectations'
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -23,7 +24,7 @@ RSpec.configure do |config|
   # assertions if you prefer.
 
   config.include FactoryGirl::Syntax::Methods
-  
+
   config.expect_with :rspec do |expectations|
     # This option will default to `true` in RSpec 4. It makes the `description`
     # and `failure_message` of custom matchers include text for helper methods
